set up docker-compose to work on windows
the "catch-all" ip-address 0.0.0.0 does not work on windows, and jekyll
requires 0.0.0.0 to serve locally. used the workaround described in this
blogpost https://tonyho.net/jekyll-docker-windows-and-0-0-0-0/ to get
docker-compose to work locally.

it boils down to running jekyll in a non-default environment by setting the
JEKYLL_ENV variable, and setting up a configuration file for this environment
that makes jekyll create links to a different location than where it thinks it
is hosted.