Closes #2: Use jekyll base image to run in local environment
The Jekyll base image is good enough for now. The first time it updates/installs the ruby gems needed to use the jekyll theme, but this is specific to the playbook solution as it is. A possible improvement to this is to build our own image based on jekyll/jekyll and do the needed gem updates so it's ready to run, but I think this is overkill at this time.